fix: ajustes no projeto

+++ src/main/java/edu/senac/service/AlunoService.java
@@ -3,6 +3,7 @@ package edu.senac.service;
 import edu.senac.dto.AlunoRequest;
 import edu.senac.dto.AlunoResponse;
 import edu.senac.entity.Aluno;
+import edu.senac.exception.AlunoException;
 import edu.senac.repository.AlunoRepository;
 import edu.senac.utils.AlunoConverter;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,27 +15,38 @@ import java.util.stream.Collectors;
 @Service
 public class AlunoService {
 
+    private static final String ALUNO_NAO_ENCONTRADO = "Aluno não encontrado";
+
     @Autowired
     private AlunoRepository alunoRepository;
 
-    public AlunoResponse criarOuAtualizarAluno(AlunoRequest alunoRequest) {
-        // TODO verificar se o aluno existe na atualização
+    public AlunoResponse criarAluno(AlunoRequest alunoRequest) {
+        Aluno novoAluno = AlunoConverter.toAluno(alunoRequest);
+        return AlunoConverter.toAlunoResponse(alunoRepository.save(novoAluno));
+    }
+
+    public AlunoResponse atualizarAluno(AlunoRequest alunoRequest) {
+        if (alunoRepository.findById(alunoRequest.getIdAluno()).isEmpty())
+            throw new AlunoException(ALUNO_NAO_ENCONTRADO);
+
         Aluno novoAluno = AlunoConverter.toAluno(alunoRequest);
         return AlunoConverter.toAlunoResponse(alunoRepository.save(novoAluno));
     }
 
     public List<AlunoResponse> buscarAluno(String nome) {
-        // TODO verificar se trouxe aluno
         return alunoRepository.findAlunoByNomeContainingIgnoreCase(nome)
-                .stream().map(c -> {
-                    return AlunoConverter.toAlunoResponse(alunoRepository.save(c));
-                }).collect(Collectors.toList());
+                .stream().map(c -> AlunoConverter.toAlunoResponse(alunoRepository.save(c))).collect(Collectors.toList());
     }
 
-    public Integer excluirAluno(Integer idAluno) {
-        // TODO verificar se tem trouxe aluno
+    public AlunoResponse excluirAluno(Integer idAluno) {
+        if (alunoRepository.findById(idAluno).isEmpty())
+            throw new AlunoException(ALUNO_NAO_ENCONTRADO);
+
+        AlunoResponse alunoResponse = new AlunoResponse();
+        alunoResponse.setIdAluno(idAluno);
+
         alunoRepository.deleteById(idAluno);
-        return idAluno;
+        return alunoResponse;
     }
 
 }
